* dev:
  Convert internal linkage statics to local namespaces.
  Improve Delegate: use shared_ptr<int> as Token. Use emplace(t, u) instead of insert(pair<T, U>(t, u)) where applicable. Improve InputHandler: exploit Delegate's multi-callback for non-consuming registrations. Support Win7.
  Components are now stored as unique_ptrs. shared_ptr<Collider> is created by CollisionManager. Actor doesn't know/care about Collider::Ptr anymore.
  Define Component::Ptr. Remove enable_shared_from_this derivation for Actor.
  Use ActorID, use unique_ptr<Actor> instead of shared_ptr<Actor>, eliminate weak_ptr<Actor> from Level's public API, use plain pointers and ActorID validity checks instead.
  Reuse Component::SClone() for Colliders, remove redundant SCloneCollider(). Use Strings::ToString everywhere. Remember state of Collider Debug Shapes, and honour it on ShapeRenderer copy. Remove RenderFactory. Fix Release build configuration.

+++ LittleEngine/Source/Levels/LevelManager.cpp
@@ -23,7 +23,7 @@ namespace LittleEngine {
 	}
 
 	bool LevelManager::LoadLevel(const LevelID& levelID) {
-		if (levelID >= 0 && levelID < createdLevels.size()) {
+		if (levelID >= 0 && static_cast<size_t>(levelID) < createdLevels.size()) {
 			if (activeLevel) {
 				activeLevel->Clear();
 			}
